Refactoring legend handling
