+++ src/components/map/plugin/PluginContextMenu.js
