+++ src/map.js
@@ -4,7 +4,7 @@ import { union } from 'lodash/fp';
 import { init, config, getUserSettings } from 'd2/lib/d2';
 import { isValidUid } from 'd2/lib/uid';
 import i18n from './locales';
-import PluginMap from './components/map/PluginMap';
+import PluginMap from './components/map/plugin/PluginMap';
 import {
     mapRequest,
     getExternalLayer,
